Adding FCM token to Settings screen

+++ android/app/src/main/java/com/sdoras/petfeeder/core/utils/BindingAdapters.kt
@@ -14,7 +14,7 @@ object BindingAdapters {
             return
         }
 
-        val whole = floor(double)
+        val whole = floor(double).toInt()
         val fraction = Fraction.createFraction(double - whole)
 
         textView.text = if(whole > 0) {
